requirements.txt
openai==0.25.0
python-dotenv==0.21.0
word-count==0.1.0
tiktoken==0.1.2
pytz==2022.7
requests==2.28.1